Naive sanitization of Idea description body multiloc
Simply bypassing the sanitization in the before_validation step
doesn't seem to fix the issue of '&amp;' displaying for '&' in
shortened version of the Idea description, displayed when viewing
Idea cards in the Project show view.
This naive solution does seem to fix the issue, but may introduce
other issues in different diplays of the body text (no such issues
found as yet).