IN-532 Reworking before/after each hooks to be able to use before(:all) in tests
IN-532 #comment before(:all) (and thus let_it_be) are run before(:each). To be able to use them easily, we need to switch to the test tenant before, they are run.

IN-532 #time 0w 0d 3h 11m Reworking before/after each hooks to be able to use before(:all) in tests