IN-532 Remove unnecessary after(:each) example hook that switch the DB back to the test tenant
IN-532 #comment If necessary, the tenant is switched back as part of the transaction associated to the example (see config.use_transactional_fixtures = true).

More importantly, the db request in the hook was also causing PG::InFailedSqlTransaction errors for some tests that contains (expected) errors in DB queries.

IN-532 #time 0w 0d 2h 42m Remove unnecessary after(:each) example hook that switch the DB back to the test tenant