IN-581 Use 'distinct' to avoid input duplications
IN-581 #comment We placed 'distinct' at the level of the subquery. If it's applied directly to the input relation, this creates errors (PG::InvalidColumnReference: ERROR: for SELECT DISTINCT, ORDER BY expressions must appear in select list) because the distinct is added to the main SQL query.

See https://stackoverflow.com/questions/42846286/pginvalidcolumnreference-error-for-select-distinct-order-by-expressions-mus for more detail.

IN-581 #time 0w 2d 7h 27m Use 'distinct' to avoid input duplications