Fix sentry error tracking for backgroud jobs
sentry-rails gem catches any active_job error, sends it to sentry
server and reraises it
https://github.com/getsentry/sentry-ruby/blob/master/sentry-rails/lib/sentry/rails/active_job.rb#L36
It worked when job is run with `perform_now`.
It seems in this case que doesn't catches errors
(maybe que is not even involved, see below).

que gem also catches any error, but reraises it only if `reraise_errors`
keyword argument is passed as true
https://github.com/que-rb/que/blob/77c6b92952b821898c393239ce0e4047b17d7dae/lib/que/job_methods.rb#L65
In this case que calls `handle_error`, and had `handle_error(error)`
with `Sentry.capture_`, but it worked only for `perform_later`,
`perform_now` ignored `handle_error`.

Sometimes `reraise_errors` is true
https://github.com/que-rb/que/blob/77c6b92952b821898c393239ce0e4047b17d7dae/lib/que/job.rb#L138
but sometimes it's false
https://github.com/que-rb/que/blob/77c6b92952b821898c393239ce0e4047b17d7dae/lib/que/active_job/extensions.rb#L19
(we have this code copied to our repo).

So the solution is to pass `reraise_errors` as true.
This way we don't need `handle_error`, sentry-rails does the job for us.