Simplify que integration. Fix `perform_retries`
1. `should_retry` was `nil` by default despite of `perform_retries true`
in `ApplicationJob`. It can be checked with
`PublishRawEventToSegmentJob.should_retry`.
Because class instance variables are not shared among class hierarcy.

2. `perform_retries false` should stop jobs from being retried
according to this refactoring (moving from sidekiq to que)
https://github.com/CitizenLabDotCo/citizenlab/commit/690e09e#diff-ca180494c4437826064e9990a73787de016bab8f906026f248b42c5267c5daacL2
and this sidekiq behaviour https://stackoverflow.com/a/59754115/4862360

But in fact failed jobs were still retried.
Even when our `handle_error` method is empty (super is not called),
failed jobs were retried.
So `perform_retries false` didn't affect retries.

3. Before this refactoring,
https://github.com/CitizenLabDotCo/citizenlab/commit/d02e677a3c80b8a1a5eaa53e294ae8e8aee81142#diff-03d233c489397f38d4056281a21e93b85db7890a6f6144e800b632462bbe968bL8
jobs expired if `should_retry` is `true`.

It's also the default behaviour
https://github.com/que-rb/que/blob/61e1805deb8161e5c1502e518a9a44fa464d308c/lib/que/job_methods.rb#L129

And indeed, if job is expired, it's not retried.