Test folder publication removed by filter - failing as expected
Currently, we are simply removing truly empty parent publications,
which equates to folders with no contents.
Now, we want to remove parent publications that contain only
children that should not be visible (e.g. filtered out by area).
This test tests if such a parent publication is removed, and fails
as we have not yet implemented this functionality.