Bugfix: Don't exclude parents of visible projects
We need to ensure we get all admin_publications representing
projects that are visible to user, and to do this we need to
ensure we don't use the top_level_only (depth=0) filtered scope
as a starting point, if this parameter has been used. This means
project_publications will include projects in folders.

Need to create a test to catch this, as current specs were all
passing!