[CL-6867] Speed up password reset email
This is my 4th attempt at tackling the issue of speeding up the reset password email. After trying to

- give the campaign mailer higher priority for reset emails (not easily doable because lots of intermediate jobs in between)
- Make the ActivityJob sync instead of async (not possible because we delay it in certains parts of the app as workarounds)
- using DeliveryService.send_now (not easily possible because it needs an activity attached to it, which is created asynchronously in the ActivityJob again)

I now resort to simply moving the email out of the whole campaign engine, greatly simplifying the involved code until an email is being sent. We lose some functionalities around tracking and alike, but I doubt this is useful for a transactional email like a password reset in the first place.