IN-673 Rework zsc_tasks factory to create coherent tasks.
There is no constraint that prevents to associate categories and inputs that corresponds to different views to a task, but in practice that should not happen. we reworked the factory to make sure that by default this doesn't happen. However, it's still responsibility of the caller to make sure to provide categories or inputs that belongs to the same category when overriding attributes.