[NO-TICKET] Fix anti-pattern of including HTTParty and deleting base_uri to the class
Delegating base_uri to the class is an anti-pattern when the base uri is not constant. That's because it relies on a class instance variable to store the URI. As a result, the base uri is shared by all instances and its value is overridden by the last call to `base_uri=`.