IN-688 Fix issue where links were referring to missing nodes.
Some links had a source or target id that didn't correspond to any node in the network. It was because of "orphan nodes", nodes that do not belong to any community (yes, that's possible -- see NLP algorithm). This list of nodes returned to the front-end being computed directly from communities, those orphan nodes were not reported. But the links between them were still there.