IN-693 Fix error when destroying views or data sources
Before this fix, deleting a view or a data source would raise:
`ActiveRecord::RecordNotDestroyed: Failed to destroy the record`

This is very likely due to the bi-directional destroy dependency
between the two models, since removing the dependency from `data_sources.rb`
solves the issue (but is not OK from a functional point-of-view).
Also, bi-directional destroy dependencies are supposed to be supported
since Rails 5, so I really could not explain why this could not work.

The before_destroy callback approach seems to solve it for some
undetermined reason.