[EN-1870] Get back projects_allowed_input_topics#show
Fow not it seems this endpoint won't be used from FE because
projects_allowed_input_topics will always be cached in `included`
part of projects#show response.
But we add it because `included` should not be used directly,
it just caches data for future responses
(projects_allowed_input_topics#show in this case).

This commit partially reverts 39122df9c1 and 6bf0b77d87