[CL-163] Improve Users XLXS Export endpoint performance
This change drastically improves the user export xlxs export functionality in the admin back end. It times out when platforms have lots of users (40k for https://creamos.injuv.cl/es-CL/).

When profiling the wall time of the controller action via app_profiler (https://github.com/Shopify/app_profiler) and speedscope (https://www.speedscope.app/) you can see lots of time spend in fetching the AppConfiguration instance over and over again. This is also observable in the Rails logs. The query is executed for every user / row in the export file. Looking at the call stack, the query can be traced back to the `Frontend::UrlService`, which is called in order to include the profile URL as a column in the export. The `UrlService` uses the app configuration to determine the URL of the profile (host / schema part).

To improve performance, the following changes have been introduced:

- Instantiate only one `UrlService` for the whole export file instead of one for every single row
- In the `UrlService` class, memoize the database call (`AppConfiguration.instance`), so it can be reused after the database has been queried the first time

In my simple local benchmark setup (10k users), this brings down the response time from 7.5 seconds to 2.5 seconds, an improvement of about 300 %.