CL-598 Temporarily log Clave unica auth response
We want to temporarily log the response auth hash that Clave Unica returns to understand where we can find the RUT value, which is what we actually want to persist into a user custom field later on.
I've tried using an epic deployment before instead of doing this on a production environment, but Clave Unica seems to whitelist only certain domains. Initiating the Clave Unica verification request always failed with the error 'invalid redirect URL'.