CL-610 Extract code into a method to remove the need for a comment
From a PR review suggestion: "If a comment is needed to explain a piece of code, abstraction is needed. Using expressive variable names and extracting code to new methods are the ways to achieve abstraction during day to day coding. By doing so, the code will be easier to read."