[CL-909] Introduce a new HeaderBgUploader that does not use deprecated Tenant
When testing how the existing `AppHeaderBgUploader` works with the `HomePage`, I've noticed that the uploader is monkey-patched for EE to take into account the tenant ID for the storage directory. See `MultiTenancy::Patches::AppHeaderBgUploader` in EE for details.
The tricky part is that `HomePage` does not have a relation to tenant as `AppConfiguration` does.

As the `Tenant` model is deprecated anyway and we want to move away from it (See TEC-15 for details), I opted to introduce a new uploader `HeaderBgUploader` here. It functions exactly the same as the old `AppHeaderBgUploader`, but works with `AppConfiguration` instead of `Tenant` and therefore doesn't need the monkey-patching anymore, which makes me very happy!
I've added deprecation notes to the relevant classes, we can remove them once the data migration is complete. The uploader also has tests now.