[CL-895] Add Pin model & join table for pinned admin publications
A pin is an entry in the database that references any kind of admin publication (Project, Folder) and a page (currently only HomePage). We add a unique index, because pinning the same admin publication on one page does not make sense.

When the admin publication or page of a pin is destroyed, we will also destroy the pin itself.