Accept nested attributes for pinned_admin_publications directly instead of pins
This change makes the API a bit easier to use because the API client does not need to deal with the join model (Pin) anymore. Instead, they can directly specify a list of admin publication ids as shown in the spec, and Rails will handle the rest in the background. Cool.