CL-704 Fix the reference-distribution factory
If the custom field had more than 2 options, the distribution would include nil counts. This is now fixed by cycling through the population counts if necessary. Ideally, we would check that the population_counts size is equal to the number of options when both the custom field and the population_counts are provided explicitly, but this is not straightforward to implement with FactoryBot API.

I also added some extra validation rules to make it easier to identify invalid distributions in the future.