[No ticket] Organise seeds
This is 10% project that escalated a little. It started by me having a hard time to understand how our seeds, especially in EE works.
Before this change we used to have a single seed file with 1000 lines. It made it very difficult to understand execution flow & branching, as well as following which records get created (for which tenant) and how all of that plays together with the base templates.

This commit does not change any behaviour of the seeds, except for one thing. I've removed the `empty` seed size option, because it seems to overlap with the `empty_localhost` tenant we are creating. If I am wrong an this is actually needed, I'm happy to reintroduce it.

The seeding process now starts by calling `MultiTenancy::Seeds::Runner#execute`. A simple Ruby class that is instantiated. It contains default values, config and helper methods (random record selection, translations etc.).
The actual records are created in sub classes of `MultiTenancy::Seeds::Base`. There are many more opportunities to refactor the code, but I had to stop somewhere. The goal of the change is to bring a bit of structure to the seeds (and invite new code to follow that structure), not to fix every single offense. I also didn't want to risk breaking existing code, so the record creation itself it almost left as-is.

I like that you can now see the clear differences between the different schemas / tenants, which before was very hard!