[No ticket] Revert sorting CL2_SUPPORTED_LOCALES alphabetically
In some places of the code we have implicit dependencies of the order of locales in `CL2_SUPPORTED_LOCALES`.

One of them makes the current e2e tests fail, which is why I'm reverting my change that sorts the locales in alphabetical order. In order to fix this properly, we should think about introducing a primary locale for tenants and get rid of the implicit dependency that currently exists for the `CL2_SUPPORTED_LOCALES` array.

I think it's okay to have english be the first item and have the rest ordered alphabetically.

Places on where we depend on the order of locales are for example:
- When creating slugs, we usually just the first value in the title_multiloc translations hash (`Project#generate_slug`)
- When selecting the nearest locale fur users, we fallback to the first platform locale in alphabetical order(AppConfiguration#closest_locale_to)