DISP-150 Add BOSA FAS public keys
Add public keys explicitly to prevent `omniauth_openid_connect` to fallback to discovery mode.

The discovery mode fails because `omniauth_openid_connect` uses the issuer URI to infer the location of the .well-known endpoint (which returns the discovery document), but the endpoint is actually located at: https://iamapps-public.int.belgium.be/.well-known/openid-configuration.