CL-1092 Add basic tests for distribution serializers
Due to the polymorphic nature of reference distributions, acceptance tests for reference distributions do not longer test the exact responses (since the response is different for each distribution type). Tests just verify that the response matches the output of the serializer. Thus, I'm adding some tests for serializers to cover the exact payloads.