Fix Rails cache containing whole records instead of primitive data
which fixes a bug that can be experienced in development mode when enabling caching via `rails dev:cache`. See https://github.com/rails/rails/issues/43856 for details.

As stated in the Rails docs, the low level Rails cache should only contain primitive data structures and not whole ActiveRecord objects:

> 1.7.1 Avoid caching instances of Active Record objects
> You should avoid this pattern. Why? Because the instance could change. In production, attributes on it could differ, or the record could be deleted. And in development, it works unreliably with cache stores that reload code when you make changes.
>
> Instead, cache the ID or some other primitive data type. [...]

https://guides.rubyonrails.org/caching_with_rails.html#avoid-caching-instances-of-active-record-objects

This commit follows through with this suggestion. Where we before saved whole records in the cache, we now only cache their IDs and fetch the records from database right after so that the API of the affected methods stay the same.

The cache keys have also been changed slightly so we are not running into the issue of fetching old cache data (with whole records) where the new application code expects IDs already.