[CL-1393] Improve smart groups query performance with low level caching
This commit aims to improve the performance issues we experience for some customers that use smart groups heavily for permission checks of projects & folders.

With smart groups, we need to execute SQL queries based on the smart group rules every time we want to check whose a member of that smart group. This can get very expensive, because the rules for the smart groups are very flexible, and so are the queries generated behind the scenes. These rules are generally unoptimized, because we don't know in advance which columns a smart group is going to query (they can be set an run time in the back office).

Additionally, the front-end is requesting resources related to projects in an ineffective matter currently. When displaying projects of a folder for example, project data is fetched one by one instead of grouped in one request. Every single request triggers another query to check for permissions of the current user. This will be fixed too at a later point in time, for now we'd like to work around these slow queries by caching them.

Many of the rules are queries against the users table. We check for specific values in their email, their role, timestamp columns and similar. By leveraging Rails low level cache, we can cache the members of those smart groups instead of calculating them again and again.
We have to be smart about how to invalidate the cache. If the rules of the smart group change, we need to recalculate its members again. If user records change, we also need to invalidate the cache and recalculate again.

This logic is new and implemented in the `Rule#{cache_key_fragments}`. The implementing class is supposed to override that method with cache key fragments that make sense for that specific rule. If no cache key fragments are specified, the rule cannot be cached.

We can go further than this change here and introduce caching for even more rules. But because caching is hard and tricky, I'd like to keep the blast radius small for now and only cache those groups that are used heavily on production and generate issues for us. Following through with the less-used groups in the future should not be an issue.