[CL-1581] Add Vienna Citizen SAML SSO
This change introduces a 2nd SAML Single Sign On for Vienna.

Before, we only had

```
/auth/saml
```

which could be used by employees of the city of Vienna to authenticate.

Now, we introduce an SSO method that is similar but aimed at citizens. It's still SAML in the background, but the IdP is a different one. We will now have these auth routes available:

```
/auth/vienna_employee (used to be /auth/saml)
/auth/vienna_citizen (new)
```

We also introduce a new feature flag so we can toggle both SSOs separately from each other.

EE PR: https://github.com/CitizenLabDotCo/citizenlab-ee/pull/334

Also:

- Change order of SingleSignOnService URLs in XML file because it confuses the SAML library in Ruby
- Reintroduce StandardPortal SSO login in Frontend, which mostly reverts #1723.