[No ticket] Improve seeding performance
Seeding the database is a task that is executed regularly because it brings our database to a consistent, sane state and gives us some data to play around with. As a developer, it's nice to have a single task that completely resets the dev environment, for example when switching branches.

Making this a fast process will overall make us more productive and happier, so I looked into what we can do to speed up the process.

With the changes in this PR, the seed time improves by a factor of ~ 3.

```
$ docker-compose run --rm web bundle exec rake db:drop db:create db:schema:load; time docker-compose run web bundle exec rake db:seed

current master
    Executed in  135.33 secs
this branch
    Executed in   49.98 secs
```

- Reduce number of users, ideas and initiatives created for normal (medium) seed size.
- Create less comments
- Do not create an empty tenant by default:

We used to create an empty localhost tenant that had no records. While this is probably useful in some cases, most of the time we only use the normal `localhost` tenant that has some data in it. You can still seed the empty localhost by specifying `SEED_EMPTY_TENANT=true` env var.