CL-1599 Rework VCR configuration
The VCR configuration is global, so I moved it to spec_helper.rb.

Unfortunately, it seems that VCR does not support a convenient way to dynamically adapt the cassette library directory (for example, to save the cassettes in the corresponding engine folder). So for now, we'll just change the cassette directory in the spec files before running the tests.