[No ticket] Fix seed spec with empty localhost tenant
After merging https://github.com/CitizenLabDotCo/citizenlab-ee/pull/341 CI started to complain that the seed spec was failing, so Simon had to revert my original changes in https://github.com/CitizenLabDotCo/citizenlab-ee/pull/350.

This change now properly fixes the spec. Because we don't seed the empy localhost by default anymore, we're stubbing the ENV for the duration of the test (and also decrease the number of expected tenants to be created from the seed file).