CL-1827 Reset CurrentAttributes before/after controller tests
Some state was leaking between tests because CurrentAttributes, and in particular `Current.tenant`, were not reset between tests. The current tenant is defined by a hook in the ApplicationController, so the problem only applies to the controller tests (at least for now). Acceptance tests are not affected by this problem, probably because they use the rails executor which resets the CurrentAttributes before and after each request.