[CL-1887] Use placeholder names when Vienna citizen SSO does not return real names
Viennas SSO for citizen does not always return first names and last names by design. The fields are optional on their side, we therefore have to deal with the scenario when a citizen authenticates via SSO and we do not have any first name or last name.

First names and last names are required for the user model, which is why we generate placeholders based on the email address, which is always present. It tries to do the most meaningful thing:

Examples

```
`alice.bob@citizenlab.co` => `A B`
`alice_bob@citizenlab.co` => `A B`
`alice@citizenlab.co` => `A L` (no second word available to deduce the last name from)
```

Another option is to make this more complicated: Save the user record without a name and ask the user in the FE to fill that out after the SSO mechanism completed. Because the sign up process is complicated already and we want to go live with this soon, we agreed with Vienna on this less-optimal solution for now.