CL-1881 Rework the analytics_fact_project_statuses view
Rework the view to make it easier for us down the road.

The new view only lists the last status. The main reason for this change is that, in its current state, the analytics API does not allow us to query the last status of each project (it has a minimal 'first' aggregation method but data in each group cannot be sorted).

Another small change is that *draft* timeline projects whose last phase is complete are no longer considered complete.

We also dropped the 'created' and 'deleted' statuses. The 'created' was pretty much useless anyway (it could be inferred from other status changes). The 'deleted' status is kind of interesting to know, but including it removes deleted projects from the list of finished projects (since we only consider the last status). Since we don't show any information about deleted projects in UI, I decided to drop it for now.