CL-1462 Remove "auto_config" mechanism from Tenant
The only place in the code that relies on the automatic creation of the app configuration when a tenant is created is when tenants are created in the tests or the test setup. In other place (e.g. MultiTenancy::TenantService), the auto config is disabled.

Here, we decided to remove that mechanism that was enabled by default and move that responsibility to the tenant factories. The goal is make the code simpler and remove bit by bit references to attributes they will removed from the Tenant model.