CL-1462 Remove code that logs lifecycle changes from SideFxTenantService
Logging lifecycle changes is now the responsibility of SideFxAppConfigurationService. That's because the lifecycle info is part of the settings attribute that will be removed from Tenant.

To avoid breaking changes in communication with external services that listen to these events, logged activities (for lifecycle changes) must still be associated to the Tenant model, even if they are triggered from the SideFxAppConfigurationService. But this was actually already implemented, since every activity logged for AppConfigurations is doubled by a similar activity for the Tenant.

Also, it's tested here: back/engines/ee/multi_tenancy/spec/services/side_fx_app_configuration_spec.rb:39