CL-1462 Refactor SideFxAppConfigurationService#after_create to trigger UpdateGoogleHostJob
Queuing UpdateGoogleHostJob jobs is now the responsibility of SideFxAppConfigurationService instead of SideFxAppTenantService. This is because this job is queued only if some conditions dependent on the settings (lifecycle_stage) of the platform are met, and the access to settings through the tenant model is deprecated and will be removed soon.