[CL-2157] Fix translation string for Vienna privacy policy
CrowdIn recognizes this as a duplicate string because the defaultMessage is the same as another string, which makes it impossible to translate it to a different string.

The translations for Vienna do indeed differ though, so I want to get rid of that 'duplication detection' on CrowdIn by giving that translation a different and unique key.