[CL-2227] Rename the `content_builder` feature flag
Currently, this feature flag is only used by the front-end to determine if the editor for project descriptions should be displayed. It has no impact on the back-end routes that allow for customizing project content or on the report builder functionality, which is controlled by a separate feature flag (`report_builder`).

It's possible that a `content_builder` feature flag would be useful in the future if we ever want to turn off all content builders at once, but that's not something we need at this time, and this feature flag was simply a misnomer.