[CL-2702] Store project_id in activities
The project ID is derived automatically from the item during the construction/initialization of the job object. This step occurs in the constructor because it needs to happen prior to serializing the job. If this didn't happen, then it would not be feasible to determine the project ID for deleted items.