[CL-2779] Always use large image for project card
On the folder page at the very top, we show a project card.
The "size" prop of the card is small, so the small image version was used.
But the real size of this card in pixels is bigger than any card
on the project list. So, the small version looks blurry.
We could use a different version on the folder page, but it requires
some work and in general using different versions is very time consuming.

Apart from that, removing the small version makes processing faster.