[CL-2909] Delete bundler cache
Disable bundler cache to decrease the size of the image.

The Bundler cache is a directory that stores a copy of each gem that is installed by Bundler. When Bundler installs a gem, it downloads the gem's files and stores them in the cache directory. If the gem is already in the cache, Bundler will use the cached copy instead of downloading it again.