[CL-3184] Refactor TenantSerializer output to support new interfaces
In this commit, the output format of the TenantSerializer was updated to be compatible with the newly introduced MultiTenancy::Templates::CreateService and MultiTenancy::Templates::ApplyService. Previously, the serializer returned a serialized tenant that could be directly used with MultiTenancy::TenantTemplateService. However, now the output needs to be converted using MultiTenancy::Templates::TenantSerializer.format_for_tenant_template_service! before it can be used with the TenantTemplateService.

The updated format includes additional information about the original model identifiers, which is necessary for importing and copying template uploads stored in S3.