[CL-3184] Specify admin_publication_attributes explicitly in internal templates
Templates containing publications (folders or projects) must explicitly specify the admin publication attributes for each publication. This can be done either with nested attributes or by including the AdminPublication in the template. This is necessary because the :set_admin_publication hook in Folder and Project is disabled when loading a template, which means that the admin publication will not be created automatically.