[CL-3515] Implement Pundit::NotAuthorizedErrorWithReason
The pull request (https://github.com/varvet/pundit/pull/625) that adds a "reason" to Pundit::NotAuthorizedError was merged, but subsequently reverted prior to the release of Pundit 2.3.0. As of now, the maintainers are soliciting feedback and evaluating alternative approaches to support custom error messages. It appears increasingly unlikely that Pundit will support the "reason" API that we are currently dependent on.

So, we're just adding the "reason" feature ourselves with this commit.