[CL-3540] Remove openid_connect fork
We no longer need to explicitly specify openid_connect as a dependency because it is already a dependency of omniauth_openid_connect.

I upgraded omniauth_openid_connect to the latest version, which uses the most recent version of openid_connect. The latest version of openid_connect was refactored to use Faraday as the HTTP client, which allows us to use a Faraday middleware to address the issue we had with JWT encoding (which is why we forked openid_connect in the first place).

As a side effect, we no longer require the HTTPClient patch since we are now using Faraday as the HTTP client.