[CL-3671] Run the migration only in the public schema
Prevent Apartment from running this migration for each schema. It seems that the "que" tables are missing in some of the schemas, which leads to the following error:
  ERROR:  relation "que_jobs" does not exist

We don't need "Que" tables (or Que objects in general — triggers, or functions) in the tenant schemas. Only the tables in the public schema are used.