[CL-3518] Update Que tables to version 6
See: https://github.com/que-rb/que/blob/master/CHANGELOG.md#200beta1-2022-03-24

In the past, we used to run Que migrations for all tenants (via Apartment enhancements of db:migrate). However, we noticed some issues on production where some tenants had the tables, while others didn't (we haven't dug into why this happened). This caused migrations to newer versions of Que tables to fail, particularly when upgrading to 1.3.1. So, we decided to only run these migrations in the public schema, which is the only schema where the tables are actually used.