[CL-3518] Adapt ActiveJobQueExtension to the new implementation of perform in que 2.0.0
It's unclear why the freeze steps were removed when originally patching Que for multi-tenancy support. However, when upgrading to Que 2.0.0, we simply transposed the same logic to the new implementation of the +perform+ method. Ideally, we should investigate why the freeze steps were initially removed and determine if there is a less invasive way to support multi-tenancy.