[CL-1483] Create user that is confirmed by default in specs
Previously, the user_confirmation feature was disabled by default
in specs (but not in production). It means that
user#confirmation_required? returns false.

In this PR, we enabled it by default in specs, and so
user#confirmation_required? started to return true.

And now we need create users that are confirmed by default.