[TAN-104] Refactor MultilocService#t signature
There is no need in passing the user to MultilocService#t as it only uses the user's locale and none of the other user attributes. Passing the user to MultilocService#t forces all callers to be aware of the user object even if they don't need it.

This commit removes the user argument from MultilocService#t and replaces it with the preferred locale.