[CL-4035] Fix extension issues in test db upon running db:reset
When db:reset was run in development, it used to “upset” the test db. In practice, “upset” means here that the extensions were not installed properly in the test db. This commit fixes that issue.

The issue stemmed from a combination of factors. Two important points to keep in mind:
- Postgres installs extensions into a specific schema (which is public by default), but it only does so once per database. Therefore, if, for instance, the hstore extension is already installed and you execute the following SQL command:
 ```sql
 CREATE EXTENSION IF NOT EXISTS hstore SCHEMA shared_extensions;
 ```
 Postgres won't take any action.
- To access an extension, the schema in which it is installed must be in the search path. This is particularly important in our case because Apartment alters the search path and removes the public schema from it when switching to a tenant.

Hence, we must ensure that the extensions are installed in one of Apartment’s "persistent" schemas, which in our case is the shared_extensions schema (see config/initializers/apartment.rb). To achieve this, we had enhanced the db:create (and db:test:purge) rake task to install the extensions in the shared_extensions schema (check out lib/tasks/db_extensions.rake).

However despite the enhancements, it was still not working in some cases for the test db and the extensions were getting installed in the public schema when running db:reset. The reason lies in the fact that when we execute db:reset in development, it runs for both the development and test environments, and each step is executed twice. Our enhancement to the `db:create` task failed account for this, as it was only running for the development environment. Subsequently, when db:schema:load was executed for the test environment, the extensions were installed, but in the public schema instead.