[CL-3422] Log out from FC when user deletes their account
Previously, we deleted a user account and only then tried to redirect to
the logout url. But at the time of redirect, the user was already deleted
and so we couldn't build a FranceConnect logout URL, because it required
`id_token_hint` stored in the database record associated with the user.

Now, first, we fetch the FranceConnect logout URL and only then delete
the user account.