[TAN-380] Convert Database Schema to SQL
This commit replaces `schema.rb` with `structure.sql` because Que uses database constructs (such as triggers) that cannot be stored as Ruby code in `schema.rb`.

This change rendered the `db:extensions` rake task obsolete, so it was removed.

Additionally, we introduced the `stub_env` spec helper to fix some broken tests. These tests were stubbing `ENV` in a way that was not compatible with some logic that manipulates environment variables during the dump/load process.