[CL-4257] Do not use autoloaded code not-autoloaded code
VerificationService is autoloaded. But `all_methods` variable is
populated (via `add_method`) in not-autoloaded code.
And when VerificationService gets reloaded, `all_methods` is
again initialized with empty array [].

And when the attributes of the serializer are being defined, no method
is used in this line
https://github.com/CitizenLabDotCo/citizenlab/blob/30bb3536e6bc813a9d96a602716165ef812b111b/back/engines/commercial/verification/app/serializers/verification/web_api/v1/verification_method_serializer.rb#L7

And so, no attribute is defined in the serializer except :name.

And on the FE, empty `method_name_multiloc` is used and so nothing is
displayed.
https://github.com/CitizenLabDotCo/citizenlab/blob/30bb3536e6bc813a9d96a602716165ef812b111b/front/app/modules/commercial/id_auth0/api/verification_methods/types.ts#L14