[TAN-624] Add a manifest to template releases
When releasing tenant templates, a manifest is now generated and stored at the root of the release folder in S3. It will save us from having to scan the entire folder each time we want to list all the templates available. The manifest includes some metadata about each template, currently limited to the list of required locales.

This change will allow us to boost the performance of tenant creation in AdminHQ. Additionally, it will allow us to prevent users from attempting to create a tenant with a template that requires the wrong locales.