[TAN-1062] Remove redundant roles from the JWT token
When a user is granted folder-moderator rights, individual project-moderator roles are automatically added to the `roles` attribute for all projects within the folder. This can result in a large array of roles that can cause the JWT token to exceed the maximum size of a cookie. To mitigate this issue, this commit removes the redundant roles from the JWT token. Although there is still a possibility that the token will exceed the maximum size, this is less likely to happen.