[TAN-1779] Use axlsx gem to write files (fixes dates)
RubyXL gem does not write dates correctly:

```ruby
[1] pry(#<XlsxService>)> cell
=> #<RubyXL::Cell(1,4): "45415", datatype=nil, style_index=3>
[2] pry(#<XlsxService>)> cell.value
=> Fri, 03 May 2024 00:00:00 +0000
[3] pry(#<XlsxService>)> new_worksheet.sheet_data[i+1][j].value
=> 45415.0
[4] pry(#<XlsxService>)> new_worksheet.sheet_data[i+1][j]
=> #<RubyXL::Cell(1,4): 45415.0, datatype=nil, style_index=0>
```

In the line [3], the value is a float, not a date.

Axlsx gem does write dates correctly:

```ruby
[1] pry(#<XlsxService>)> rows[1][4]
=> #<RubyXL::Cell(2,4): "45415", datatype=nil, style_index=3>
[2] pry(#<XlsxService>)> rows[1][4].value
=> Fri, 03 May 2024 00:00:00 +0000
[3] pry(#<XlsxService>)> new_sheet.rows[1][4].value
=> Fri, 03 May 2024 00:00:00 +0000
```

In the line [3], the value is a date.

Probably, we should always use Axlsx gem to write files
and RubyXL gem to read files (Axlsx cannot read).